requirements.txt
flask==2.0.2
line-bot-sdk==1.20.0
bs4==0.0.1
requests==2.24.0
lxml==4.6.3
html5lib==1.1
youtube-search-python==1.4.9
psycopg2==2.9.1
pyshorteners==1.0.1